fix some error messages about input paths
