+++ docs/changelog.txt
@@ -1,3 +1,5 @@
+- Fixed some error messages about input paths.
+
 ver 0.6.0
 - Supported ucas/utoc assets.
 - Added support for UE5.4.
